+++ src/profile/Profile.js
@@ -1,12 +1,12 @@
-import {useEffect, useState} from "react";
+import {useContext, useEffect, useState} from "react";
 import {Host} from "../BackEnd";
 import {Items} from "../home/dashboard/items/Items";
+import {UserContext} from "../AppContext";
 
 export function Profile({
-                            id = "", userName = "", motto = "", url = " ",
-                            currentUser = "", userToken = "", onUpdate = () => {
-    }
+                            id = "", userName = "", motto = "", url = " ", onUpdate = () => {}
                         }) {
+    const {user} = useContext(UserContext)
     const [edit, setEdit] = useState(false)
     const [items, setItems] = useState({});
     function toggleEdit() {
@@ -16,7 +16,7 @@ export function Profile({
     function updateProfile(e) {
         e.preventDefault()
         const data = JSON.stringify(Object.fromEntries(new FormData(e.target)))
-        let bearerToken = "Bearer " + userToken
+        let bearerToken = "Bearer " + user.token
         fetch(Host + "/profile?id=" + id, {
             method: "PUT",
             headers: {"Authorization": bearerToken},
@@ -51,11 +51,11 @@ export function Profile({
                 <div className={"Profile"}>
                     <h1>@{id}</h1>
                     <h2>{userName}</h2>
-                    {currentUser === id ?
+                    {user.user === id ?
                         <button className={"Button EditProfile"} onClick={toggleEdit}>Edit Profile</button> : ""}
                 </div>
                 <hr/>
-                <Items items={items} user={{name:currentUser, token: userToken}}/>
+                <Items items={items}/>
             </>
             : <div>
                 <form className={"ProfileEditForm"} onSubmit={updateProfile}>