+++ changelog.md
@@ -2,6 +2,11 @@
 ### Fixed
 * Show search bar on smaller phones
 * Check if session is active when reloading
+* Smaller search on Desktop
+
+##  [v1.1.2] - _2023 00 00_
+### Refactor
+* moved user to context
 
 ##  [v1.1.1] - _2022 12 30_
 ### Fixed